--- src/main/java/myBossException.java
@@ -1,5 +1,0 @@
-public class myBossException extends Exception {
-    public myBossException(String errorMessage) {
-        super(errorMessage);
-    }
-}
