+++ src/main/java/myboss/Event.java
@@ -1,3 +1,5 @@
+package myboss;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
