+++ src/main/java/myboss/MyBossException.java
@@ -1,0 +1,7 @@
+package myboss;
+
+public class MyBossException extends Exception {
+    public MyBossException(String errorMessage) {
+        super(errorMessage);
+    }
+}
