+++ src/main/java/myboss/Storage.java
@@ -1,3 +1,5 @@
+package myboss;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
