+++ src/main/java/myboss/Ui.java
@@ -1,3 +1,5 @@
+package myboss;
+
 import java.util.Scanner;
 
 import java.util.ArrayList;
